add relation and validation

+++ database/migrations/2023_06_10_071301_create_customers_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->string('adress');
-            $table->integer('nip');
+            $table->unsignedBigInteger('nip');
             $table->timestamps();
         });
     }
